# backend/swagger-export.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/whatsapp/send": {
      "post": {
        "operationId": "WhatsAppController_sendMessage",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Send WhatsApp message",
        "tags": [
          "WhatsApp"
        ]
      }
    },
...
    "/whatsapp/webhook": {
      "get": {
        "operationId": "WhatsAppController_verifyWebhook",
        "parameters": [
          {
            "name": "hub.mode",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hub.verify_token",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hub.challenge",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Verify WhatsApp webhook",
        "tags": [
          "WhatsApp"
        ]
      },
      "post": {
        "operationId": "WhatsAppController_handleWebhook",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Handle WhatsApp webhook",
        "tags": [
          "WhatsApp"
        ]
      }
    },
    "/whatsapp/messages/{leadId}": {
      "get": {
        "operationId": "WhatsAppController_getMessages",
        "parameters": [
          {
            "name": "leadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get WhatsApp messages for a lead",
        "tags": [
          "WhatsApp"
        ]
      }
    },
    "/whatsapp/conversations": {
      "get": {
        "operationId": "WhatsAppController_getConversations",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get WhatsApp conversations",
        "tags": [
          "WhatsApp"
        ]
      }
    },
    "/whatsapp/analytics": {
      "get": {
        "operationId": "WhatsAppController_getAnalytics",
        "parameters": [
          {
            "name": "timeRange",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get WhatsApp analytics",
        "tags": [
          "WhatsApp"
        ]
      }
    },
...
    "/campaigns": {
      "post": {
        "operationId": "CampaignsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCampaignDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new campaign",
        "tags": [
          "Campaigns"
        ]
      },
      "get": {
        "operationId": "CampaignsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "Active",
                "Paused",
                "Completed",
                "Archived"
              ]
            }
          },
          {
            "name": "priority",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "High",
                "Medium",
                "Low"
              ]
            }
          },
          {
            "name": "assigneeId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List campaigns",
        "tags": [
          "Campaigns"
        ]
      }
    },
    "/campaigns/{id}": {
      "get": {
        "operationId": "CampaignsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get campaign details",
        "tags": [
          "Campaigns"
        ]
      },
      "patch": {
        "operationId": "CampaignsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCampaignDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update campaign",
        "tags": [
          "Campaigns"
        ]
      },
      "delete": {
        "operationId": "CampaignsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Close/Delete campaign",
        "tags": [
          "Campaigns"
        ]
      }
    },
...
    "/campaigns/{id}/run": {
      "post": {
        "operationId": "CampaignsController_runCampaign",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Run a bulk messaging campaign",
        "tags": [
          "Campaigns"
        ]
      }
    },
...
    "/templates": {
      "get": {
        "operationId": "TemplatesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Templates"
        ]
      },
      "post": {
        "operationId": "TemplatesController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Templates"
        ]
      }
    },
...
  },
  "info": {
    "title": "Telecalling CRM API",
    "description": "Scalable REST API for the Telecalling CRM platform",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateCampaignDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "@facebook-ad-leads"
          },
          "priority": {
            "type": "string",
            "enum": [
              "High",
              "Medium",
              "Low"
            ],
            "default": "Medium"
          },
          "status": {
            "type": "string",
            "enum": [
              "Active",
              "Paused",
              "Completed",
              "Archived"
            ],
            "default": "Active"
          },
          "assigneeIds": {
            "description": "Array of user IDs assigned to this campaign",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "description": {
            "type": "string",
            "example": "Campaign description"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateCampaignDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "@facebook-ad-leads"
          },
          "priority": {
            "type": "string",
            "enum": [
              "High",
              "Medium",
              "Low"
            ],
            "default": "Medium"
          },
          "status": {
            "type": "string",
            "enum": [
              "Active",
              "Paused",
              "Completed",
              "Archived"
            ],
            "default": "Active"
          },
          "assigneeIds": {
            "description": "Array of user IDs assigned to this campaign",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "description": {
            "type": "string",
            "example": "Campaign description"
          }
        }
      },
...
    }
  }
}